working!

+++ app/src/components/Program.vue
@@ -1,21 +1,45 @@
 <template>
-  <article>
-    <h3>{{ program.title }}</h3>
-    <p><strong>Host:</strong> {{ program.host }}</p>
-    <p><strong>First aired:</strong> {{ program.yearStarted }}</p>
-    <p><strong>Airs daily?</strong> {{ airsDaily }}
-    <p><strong>Audience Size:</strong> {{ program.audienceSize }}</p>
-    <p><strong>Genre:</strong> {{ program.genre }}</p>
-    <p><strong>Description:</strong> {{program.description}}</p>
-    <button v-on:click="handleDelete">delete</button>
-  </article>
+  <div>
+    <article v-if="!showEditForm">
+      <h3>{{ program.title }}</h3>
+      <p><strong>Host:</strong> {{ program.host }}</p>
+      <p><strong>First aired:</strong> {{ program.yearStarted }}</p>
+      <p><strong>Airs daily?</strong> {{ airsDaily }}
+      <p><strong>Audience Size:</strong> {{ program.audienceSize }}</p>
+      <p><strong>Genre:</strong> {{ program.genre }}</p>
+      <p><strong>Description:</strong> {{program.description}}</p>
+      <button v-if=!showEditForm v-on:click="showEditForm = true">edit</button>
+      <button v-on:click="handleDelete">delete</button>
+    </article>
+    <ProgramForm 
+        v-if="showEditForm" 
+        label="Update"
+        :program="program" 
+        :onFormSubmit="handleUpdate"
+        :onCancelEdit="() => showEditForm = false"
+    />
+  </div>
 </template>
 
 <script>
+import ProgramForm from './ProgramForm.vue';
+
 export default {
+  data() {
+    return {
+      showEditForm: false
+    };
+  },
+  components: {
+    ProgramForm
+  },
   props: {
     program: Object,
     onDelete: {
+      type: Function,
+      required: true
+    },
+    onUpdate: {
       type: Function,
       required: true
     }
@@ -27,7 +51,15 @@ export default {
   },
   methods: {
     handleDelete() {
-      this.onDelete(this.program);
+      if(confirm(`Are you sure you want to remove ${this.program.title}?`)) {
+        return this.onDelete(this.program)
+      }
+    },
+    handleUpdate(editedProgram) {
+      return this.onUpdate(editedProgram)
+        .then (() => {
+          this.showEditForm = false;
+        });
     }
   }
 };
